build: Add autotools and flatpak wrappers
This will be the way to build the app using flatpak, setuptools can
be used to build the app in a virtualenv if required.

Note that the implementation here is particularly horrible - I'm
planning to use some packages from pypi to do the webserver side
of things and integrating the way pypi wants a package to be installed
with the way that autotools wants a package to be installed is
particularly horrible.

We can't just use pip install directly, since /usr is read-only
from within flatpak. So instead we create a virtualenv, install
everything to there, then relocate the virtualenv to /app/share/eos-companion-service
and set PYTHONPATH etc and hope for the best. Its really really
horrible, I know. Suggestions welcome.