eoscompanion: Add list_applications endpoint
For now we have some very not-ideal things going on.

(1) For some reason passing a GPtrArray to PyGI with transfer full
causes it to immediately unref all the elements, which crashes
when we attempt to use it later.
(2) We're using a whitelist
(3) We're relying on app names being encoded in a certain way
(4) We're assuming that the desktop filename for an app corresponds
to its app id.

https://phabricator.endlessm.com/T20090