eoscompanion: Add middleware to ensure application is held during requests
Otherwise it is possible that the application could go away whilst we're
in the middle of handling a request.

An unfortunate side effect of this is that since we are running in
service mode, holding and then releasing the application will cause
the hold count to go to zero, which means that the application goes
away as soon as it handles the first request.