eoscompanion: Add bytes_to_string and string_to_bytes helpers
We're the helpers that we have to get data out of the shards read it
out as bytes. There isn't a way to convert a GBytes to a string, which
is understandable since a GBytes holds a superset of possible data
than just strings. However, if we know it is a string, we need to
convert it back to that so that we can parse it on the Python
side and work with it. Hence the reason for these helpers.

Unfortunately the operation of these helpers result in a lot of
copying. It'd be nice if there was a way to elide those copies, but
using a size outparam with the the array-length annotation resulted
in a crash deep in python itself.