helper: Add functions to seek an EosShardBlob
In some cases, we can just seek O(1) by using GSeekable's seek()
method. However, if the blob is compressed, then we'll need to
do it the O(N) way by reading the blob up until the desired
seek point and then returning the input stream to the
uncompressing blob for the client to deal with, using
g_input_stream_skip.