eoscompanion: Implement real content streaming
This replaces the old /content_data endpoint with a
/content_stream endpoint which supports the 'Range'
header documented here: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Range

Note that a certain popular web-browser is extra stringent on what
it expects header values to be, so pay attention to the comments
in the function before modifying it.

This should make it possible to seek video files.

https://phabricator.endlessm.com/T20380