eoscompanion: Wrap and modify stream before taking its content size
We accidentally dropped this when supporting streaming content. Examine
the blob's content type and if necessary, read the stream into memory
and modify it in place, wrapping it in a GMemoryInputStream before
passing it back to the caller such that it can be spliced in place. This
also allows us to adjust the size of the Content-Length header to
account for the change in the stream size due to modification.

https://phabricator.endlessm.com/T20732