eoscompanion: Add all_asynchronous_function_calls_closure
This is a helper closure to manage the execution of several
asynchronous function calls running in parallel. It is similar
to JavaScript's Promise.all() function, in that it collects all
the results from the asynchronous function calls into an array
guaranteed to be in the same order as the provided callbacks. Once
all the callbacks have completed, the array of results is passed
to the done_callback function for further processing.