eoscompanion: Take an Inhibit lock on idle
This will prevent the system from automatically going to sleep
or shutting down whilst the Companion App Service is running. The
service now inserts a new header, X-Endless-Alive-For-Further,
into all responses indicating the number of milliseconds from
the completion of the response that the server is guaranteed,
provided no user intervention or power management failures to stick
around for.

The lock will be automatically released when the process goes
away, since fds are automatically closed when processes exit.

A new policykit rule was also added to explictly permit
the companion-app-helper to inhibit suspend.

https:/phabricator.endlessm.com/T20963