eoscompanion: Query at least limit + offset results from apps
The per-source limit when the global limit is present is really
an optimization to prevent fetching too much data from each source
when a cap will be applied, however, we were appying this optimization
incorrectly - the limit and offset essentially function as a slice
into the global sorted results list, so for each source we need
to guarantee that we're going to get at query as many results
as would fit within that slice.

Previously we set the per-source limit to the same value as the
global limit but this is actually incorrect. Say for instance, we
had a limit of 1 and an offset of 1. In order for that limit and
offset to be applied correctly, we would actually need at least
two results from one source. If we only limited it to one, the
slice would overrun the returned list length and we would get
back no results, even though there were some in the database that
we could have easily used.