polkit: Allow inhibit-block-shutdown too
For some reason, if the service is running through systemd socket activation
and not as a regular user, it requests the inhibit-block-shutdown permission
instead of inhibit-block-sleep, regardless of the fact that we asked for
a sleep inhibitor.

It isn't clear why this is, but if we allow the shutdown inhibitor on
the polkit side, everything works.

https://phabricator.endlessm.com/T20963