fixup! build_app: Simulate what flatpak does, don't use it
Using flatpak directly means that we would have to include
flatpak in the build whilst running the bundle- job if we want tests
to run. That would be far too painful due to flatpak's rather long
dependency chain.

Really, we don't care so much about the fake apps being flatpaks
as much as we care about them having the right directory structure,
so just simulate what flatpak does with respect to exports.