test_data: Check in copies of the generated gresource and shardfiles
Having to do this is absolutely not ideal, but it saves us from
breakage down the line if basin or any other part of the stack
breaks compatibility and breaks our tests.

Ideally we should try and abstract away the dependency on EknContent
so that it can be mocked out.