service: Break the ekncontent dependency and use EknServices to get data
Previously we read content data from ekncontent directly. Amongst
other things, this was bad because it tied us to a single SDK version
because we can't load more than one ekncontent library at a time. This
commit breaks that dependency by using EknServices to perform queries
instead. It uses the ArticleMetadata query interface in the corresponding
pull rquest (https://github.com/endlessm/eos-knowledge-services/pull/38)
to perform queries in a similar way that we performed queries against
eknservices directly, except that data is marshalled through D-Bus as
opposed to a binary interface.

In anticipation of later commits that will need to mock out
this later, a new class was created in eknservices_bridge
called EknServicesContentDbConnection which is responsible for
running the queries and deserializing the response into a python
object.

https://phabricator.endlessm.com/T20853