routes: Add /resource route handler
This is an internal route (no metric is attached) which can be
used to load URI encoded internal resources. Typically this
would be used on pages where some pre-rendering has occurred
beforehand which rewrote resource:/// style URIs into
server-relative URIs (/resource), where the resource:/// or file://
URI is passed as a querystring parameter to /resource. It fetches
the resource asynchronously and then returns it as appropriate.

https://phabricator.endlessm.com/T21272