adjusters: Also render a wrapper template
This is similar to the "article-wrapper.mst" that articleHTMLRenderer
uses, but it is our own fork of that code with a few important differences.

First, we remove a bunch of stuff from our own template that isn't
useful on mobile (copy button, share buttons).

Second, the crosslinks computation has some extra steps to resolve
the URIs to be relative to the Companion App Service as opposed to
just being ekn:/// style URIs.

Third, the set-membership computation needs to happen at query
time, so we have to go out to the database we're in to work out
all the sets in our app ID.

https://phabricator.endlessm.com/T21272