flatpak: Use a fork of basin that supports "path" for now
The "source" entry in the db.json file was actually meant to indicate,
generally, where a file came from so that the renderer could use that
to pick the right rendering backend. Unfortunately basin overloaded
this to mean the path to the file on disk.

We're going to eventually get rid of our usage of basin in
the Companion App Service' tests, so just use a fork for now
to get around this problem.