List applications endpoint: blacklist com.endlessm.encyclopedia.*
The encyclopedia content is not structured to be displayed
in the categories view, nor can we display all the available
articles in one list.

Until the app is restructured blacklist it from the
application list.

The encyclopedia articles can still be accessed from
the search.

Fix for a713ef01caf48a4b5b9db9dd7709ad4a1d57ba30

https://phabricator.endlessm.com/T21042