adjuster: Add EknContentAdjuster class
This is an implementation of a more generic ContentAdjuster
type of class, which encapsulates some state needed to adjust
content, as well as a function used to determine whether a particular
content type needs to have its content adjusted (performance
optimization) and a render_async function to asynchronously
perform rendering operations.

This class encapsulates all the rendering operations done on
Ekn content.