routes: Make sure to emit 'finished' once splicing is done.
libsoup will not do that for us, since we already stole
the output stream and for all libsoup is concerned, msg
is a dead object.

However, in implementing this fix we exposed a rather unfortunate
interaction between the python garbage collector and libsoup. libsoup
really expects the object to be finalized as soon as steal_connection
is called. In expectation of finalization in that case, sockets
are closed in the finalize handler. However, with language bindings
the object sticks around in a dead state.

Referencing msg from an inner function in python as part of the
outer function's closure creates a cyclic reference, so python
will never get rid of msg. Because python never gets rid of it
the sockets never get closed and we eventually stop responding
to connections.

For now, if we just set it to None, that helps the garbage
collector figure out that it needs to get rid of it.

https://phabricator.endlessm.com/T22307