routes: Use explicit enum error codes and domains
As opposed to just passing through the error code and domain
from the GLib.Error. We depend on the enum information in
serialize_error_to_json_object, however this is not preserved
across language boundaries, which means that serialize_error_to_json
object will fail if we try to pass the domain and code directly. As
such, we need to pass explicit error codes for what we went wrong.