routes: Handle the fact that load_application_info failures are now errors
Previously we could return NULL for the GDesktopAppInfo and that
would not be a failure (the caller had to handle that case). Now, however
it is a failure since we are returning more than the GDesktopAppInfo
(we are also returning the EknServices version to use as well). Instead
of having the caller have to deal with a half-constructed object, just
make a failure to load any part an error.