tmpfiles: Also create /var/lib/eos-companion-app/.local/share
We bind-mount these already in the EknServicesMultiplexer flatpak
but the bind mount won't be created unless the directory already
exists. If it can't be created, it will silently fail, reverting
back to an overlay fs which is private to the namespace of the
flatpak itself, which will mean that xapian-bridge is unable
to find all the directories and symlinks we helpfully
created for it.

https://phabricator.endlessm.com/T22393