helper: Use gtk_icon_info_load_icon_async, as opposed to threads
gtk_icon_theme_load_icon_data is not thread safe as it
mutates an internal cache both when the pixbuf is unreffed and
when an icon is loaded. Instead of putting all that behind
a thread, look up the icon info in the cache first on the main
thread and then load the icon data asynchronously.

https://phabricator.endlessm.com/T22584