responses: Add respond_if_error_set and error response helpers
The latter saves some typing when making a json response with an
error.

The former is a helper that can be used in the following pattern:

    if respond_if_error_set(msg, error):
        ...cleanup...
        return

It returns True if the passed error is not None and internally
translates the passed GLib.Error into the corresponding
EosCompanionAppService.Error then calls error_response with it. The
caller can pass any details about the call that the client might find
useful. The caller can also pass an error_mappings dict to specify how
errors from other domains are translated into
EosCompanionAppService.Error. If no mapping exists it will be translated
into EosCompanionAppService.Error.FAILED.

The respond_if_error_set will be extended later to translate
Gio.Error.CANCELLED into the corresponding EosCompanionAppService.Error