helper: Add EosCompanionAppServiceManagedCache
This class is a GObject wrapper around a threadsafe key-value store
to be passed into functions that need to cache things such as
metadata file contents. It only exposes a single introspectable function,
clear(), which can be used to drop all the caches such that values
need to be read from disk again. This will be used later to ensure that
we drop caches when the Flatpak installation state changes.

The non-introspectable API includes lock(key), which "checks out"
and mutex locks a cache for a given key (blocking other threads
until it is unlocked) and unlock(key) which unlocks the cache, allowing
other threads to lock it.